Conflicts:
	ct_optcon/include/ct/optcon/costfunction/CostFunctionQuadraticSimple.hpp
	ct_optcon/include/ct/optcon/dms/dms_core/DmsSolver.h
	ct_optcon/include/ct/optcon/nlp/OptVector.h
	ct_optcon/include/ct/optcon/solver/OptConSolver.h

+++ ct_optcon/include/ct/optcon/nlp/DiscreteConstraintContainerBase.h
@@ -110,7 +110,7 @@ public:
 			ind += cSize;
 		}
 
-		assert(ind == c_nlp.rows()); // or throw an error
+		assert(static_cast<int>(ind) == c_nlp.rows()); // or throw an error
 	}
 
 	/**
@@ -132,7 +132,7 @@ public:
 			ind += nnEle;
 		}
 
-		assert(ind == (size_t) nzz_jac_g);
+		assert(static_cast<int>(ind) == nzz_jac_g);
 	}
 
 	/**
@@ -163,7 +163,7 @@ public:
 			ind += nnEle;
 		}
 
-		assert(ind == (size_t) nnz_jac_g);
+		assert(static_cast<int>(ind) == nnz_jac_g);
 	}
 
 	/**
